Extract models from model.ts
