+++ src/components/App.tsx
@@ -48,7 +48,7 @@ import {
   setViewType,
   logLn,
 } from "../actions/AppActions";
-import { Project, File, FileType, Directory, shallowCompare, ModelRef, fileTypeForExtension } from "../model";
+import { Project, File, FileType, Directory, ModelRef } from "../models";
 import { Service, Language } from "../service";
 import { Split, SplitOrientation, SplitInfo } from "./Split";
