+++ src/components/ControlCenter.tsx
@@ -25,8 +25,7 @@ import { EditorView, View, Tab, Tabs } from "./editor";
 import { Sandbox } from "./Sandbox";
 import { GoThreeBars, GoFile } from "./shared/Icons";
 import { Button } from "./shared/Button";
-import { FileType, getIconForFileType, Problem, ModelRef } from "../model";
-import { Project, File, Directory, shallowCompare } from "../model";
+import { File } from "../models";
 import { Problems } from "./Problems";
 import appStore from "../stores/AppStore";
 import { layout } from "../util";
