+++ src/components/NewProjectDialog.tsx
@@ -24,7 +24,6 @@ import { Service, IFiddleFile } from "../service";
 import * as ReactModal from "react-modal";
 import { Button } from "./shared/Button";
 import { GoGear, GoFile, GoX, Icon } from "./shared/Icons";
-import { File, FileType, Directory, extensionForFileType, nameForFileType, Project } from "../model";
 import { KeyboardEvent, ChangeEvent, ChangeEventHandler } from "react";
 import { ListBox, ListItem, TextInputBox } from "./Widgets";
 import fetchTemplates from "../utils/fetchTemplates";
