+++ src/index.tsx
@@ -23,7 +23,6 @@ import * as React from "react";
 import * as ReactDOM from "react-dom";
 import { Workspace } from "./components/Workspace";
 import { EditorView, ViewTabs, Tab, Tabs } from "./components/editor";
-import { Project } from "./model";
 import { Header } from "./components/Header";
 import { Split } from "./components/Split";
 import { Toolbar } from "./components/Toolbar";
