+++ src/stores/AppStore.ts
@@ -19,7 +19,7 @@
  * SOFTWARE.
  */
 
-import { EventDispatcher, ModelRef, Project, File, Directory, FileType, SandboxRun } from "../model";
+import { EventDispatcher, ModelRef, Project, File, Directory, FileType, SandboxRun } from "../models";
 import { Service } from "../service";
 
 import dispatcher from "../dispatcher";
