+++ tests/unit/Project.spec.ts
@@ -1,7 +1,7 @@
 /* Any copyright is dedicated to the Public Domain.
  * http://creativecommons.org/publicdomain/zero/1.0/ */
 
-import { Project } from "../../src/model";
+import { Project } from "../../src/models";
 
 describe("Project tests", () => {
   describe("constructor", () => {
