+++ tests/unit/directory.spec.ts
@@ -1,7 +1,7 @@
 /* Any copyright is dedicated to the Public Domain.
  * http://creativecommons.org/publicdomain/zero/1.0/ */
 
-import { Directory, FileType, File } from "../../src/model";
+import { Directory, FileType, File } from "../../src/models";
 
 function getDirectoryStructure() {
   const a = new Directory("test");
